requirements.txt
langdetect
Pillow
protobuf
streamlit
google-generativeai
google-genai
